requirements.txt
django
gunicorn
django-heroku
djangorestframework
django-cors-headers
django-allauth
django-rest-auth
Pillow
djongo
dnspython
djangorestframework-jwt
django-allauth
djangorestframework-simplejwt



